Remove ApplicationServices project
Move into Api project.
