+++ src/api/EmployerFavouritesApi/Domain/IEmployerAccountRepository.cs
@@ -1,0 +1,9 @@
+using System.Threading.Tasks;
+
+namespace DfE.EmployerFavourites.Api.Domain
+{
+    public interface IEmployerAccountRepository
+    {
+        Task<string> GetEmployerAccountId(string userId);
+    }
+}
