+++ src/api/EmployerFavouritesApi/Domain/IFavouritesReadRepository.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+using DfE.EmployerFavourites.Api.Domain.ReadModel;
+
+namespace DfE.EmployerFavourites.Api.Domain
+{
+    public interface IFavouritesReadRepository
+    {
+        Task<ApprenticeshipFavourites> GetApprenticeshipFavourites(string employerAccountId);
+    }
+}
