+++ src/api/EmployerFavouritesApi/Domain/IFavouritesWriteRepository.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+using DfE.EmployerFavourites.Api.Domain.WriteModel;
+
+namespace DfE.EmployerFavourites.Api.Domain
+{
+    public interface IFavouritesWriteRepository
+    {
+        Task SaveApprenticeshipFavourites(string employerAccountId, ApprenticeshipFavourites apprenticeshipFavourite);
+    }
+}
