+++ src/api/EmployerFavouritesApi/Domain/ReadModel/ApprenticeshipFavourites.cs
@@ -1,0 +1,8 @@
+using System.Collections.Generic;
+
+namespace DfE.EmployerFavourites.Api.Domain.ReadModel
+{
+    public class ApprenticeshipFavourites : List<ApprenticeshipFavourite>
+    {
+    }
+}
