+++ src/api/EmployerFavouritesApi/Domain/ReadModel/Provider.cs
@@ -1,0 +1,8 @@
+﻿namespace DfE.EmployerFavourites.Api.Domain.ReadModel
+{
+    public class Provider
+    {
+        public int Ukprn { get; set; }
+        public string Name { get; set; }
+    }
+}
