+++ src/api/EmployerFavouritesApi/Domain/WriteModel/ApprenticeshipFavourite.cs
@@ -1,0 +1,15 @@
+using System.Collections.Generic;
+
+namespace DfE.EmployerFavourites.Api.Domain.WriteModel
+{
+    public class ApprenticeshipFavourite
+    {
+        public ApprenticeshipFavourite()
+        {
+            Ukprns = new List<int>();
+        }
+
+        public string ApprenticeshipId { get; set; }
+        public IList<int> Ukprns { get; set; }
+    }
+}
