+++ src/api/EmployerFavouritesApi/Domain/WriteModel/ApprenticeshipFavourites.cs
@@ -1,0 +1,8 @@
+using System.Collections.Generic;
+
+namespace DfE.EmployerFavourites.Api.Domain.WriteModel
+{
+    public class ApprenticeshipFavourites : List<ApprenticeshipFavourite>
+    {
+    }
+}
