+++ src/api/EmployerFavouritesApi/Infrastructure/Configuration/ConnectionStrings.cs
@@ -1,0 +1,8 @@
+namespace DfE.EmployerFavourites.Api.Infrastructure.Configuration
+{
+    public class ConnectionStrings
+    {
+        public string Redis { get; set; }
+        public string AzureStorage { get; set; }
+    }
+}
