+++ src/api/EmployerFavouritesApi/Infrastructure/Interfaces/IFatRepository.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+
+namespace DfE.EmployerFavourites.Api.Infrastructure.Interfaces
+{
+    public interface IFatRepository
+    {
+        Task<string> GetApprenticeshipNameAsync(string apprenticeshipId);
+        Task<string> GetProviderNameAsync(int ukprn);
+    }
+}
