+++ src/api/EmployerFavouritesApi/Startup.cs
@@ -3,12 +3,11 @@ using System.IO;
 using System.Reflection;
 using DfE.EmployerFavourites.Api.Helpers;
 using DfE.EmployerFavourites.Api.Security;
-using DfE.EmployerFavourites.ApplicationServices.Commands;
-using DfE.EmployerFavourites.ApplicationServices.Configuration;
-using DfE.EmployerFavourites.ApplicationServices.Domain;
-using DfE.EmployerFavourites.ApplicationServices.Infrastructure;
-using DfE.EmployerFavourites.ApplicationServices.Infrastructure.Interfaces;
-using DfE.EmployerFavourites.ApplicationServices.Queries;
+using DfE.EmployerFavourites.Api.Infrastructure.Configuration;
+using DfE.EmployerFavourites.Api.Domain;
+using DfE.EmployerFavourites.Api.Infrastructure;
+using DfE.EmployerFavourites.Api.Infrastructure.Interfaces;
+using DfE.EmployerFavourites.Api.Application.Queries;
 using MediatR;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
