+++ src/api/IntegrationTests/GeneralTests.cs
@@ -1,0 +1,47 @@
+using System.IO;
+using System.Threading.Tasks;
+using DfE.EmployerFavourites.Api;
+using Microsoft.AspNetCore.Mvc.Authorization;
+using Microsoft.AspNetCore.Mvc.Testing;
+using Microsoft.AspNetCore.TestHost;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Xunit;
+
+namespace DfE.EmployerFavourites.IntegrationTests
+{
+    public class GeneralTests : IClassFixture<WebApplicationFactory<Startup>>
+    {
+        private readonly WebApplicationFactory<Startup> _factory;
+
+        public GeneralTests(WebApplicationFactory<Startup> factory)
+        {
+            _factory = factory.WithWebHostBuilder(builder =>
+            {
+                builder.ConfigureAppConfiguration((hostingContext, config) =>
+                {
+                    config.Sources.Clear();
+
+                    config.SetBasePath(Directory.GetCurrentDirectory());
+                    config.AddJsonFile("appSettings.json", optional: false, reloadOnChange: false);
+                });
+            });
+        }
+
+        private System.Net.Http.HttpClient BuildClient()
+        {
+            return _factory.WithWebHostBuilder(builder =>
+            {
+                builder.ConfigureTestServices(services =>
+                {
+                    services.AddMvc(options =>
+                    {
+                        options.Filters.Add(new AllowAnonymousFilter());
+                    });
+                   
+                });
+            })
+            .CreateClient();
+        }
+    }
+}
