+++ src/api/IntegrationTests/appSettings.json
@@ -1,0 +1,2 @@
+{
+}
