+++ src/api/favourites-api.sln
@@ -1,0 +1,62 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26124.0
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmployerFavouritesApi", "EmployerFavouritesApi\EmployerFavouritesApi.csproj", "{2C8D96D9-A7CA-4E1C-83E0-D0924DD4A307}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IntegrationTests", "IntegrationTests\IntegrationTests.csproj", "{0DCEB08E-BA94-4968-AD0D-9504ACAA951F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitTests", "UnitTests\UnitTests.csproj", "{7282D897-351B-4F69-BF88-CDA3FC85E374}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{2C8D96D9-A7CA-4E1C-83E0-D0924DD4A307}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2C8D96D9-A7CA-4E1C-83E0-D0924DD4A307}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2C8D96D9-A7CA-4E1C-83E0-D0924DD4A307}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{2C8D96D9-A7CA-4E1C-83E0-D0924DD4A307}.Debug|x64.Build.0 = Debug|Any CPU
+		{2C8D96D9-A7CA-4E1C-83E0-D0924DD4A307}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{2C8D96D9-A7CA-4E1C-83E0-D0924DD4A307}.Debug|x86.Build.0 = Debug|Any CPU
+		{2C8D96D9-A7CA-4E1C-83E0-D0924DD4A307}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2C8D96D9-A7CA-4E1C-83E0-D0924DD4A307}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2C8D96D9-A7CA-4E1C-83E0-D0924DD4A307}.Release|x64.ActiveCfg = Release|Any CPU
+		{2C8D96D9-A7CA-4E1C-83E0-D0924DD4A307}.Release|x64.Build.0 = Release|Any CPU
+		{2C8D96D9-A7CA-4E1C-83E0-D0924DD4A307}.Release|x86.ActiveCfg = Release|Any CPU
+		{2C8D96D9-A7CA-4E1C-83E0-D0924DD4A307}.Release|x86.Build.0 = Release|Any CPU
+		{0DCEB08E-BA94-4968-AD0D-9504ACAA951F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0DCEB08E-BA94-4968-AD0D-9504ACAA951F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0DCEB08E-BA94-4968-AD0D-9504ACAA951F}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{0DCEB08E-BA94-4968-AD0D-9504ACAA951F}.Debug|x64.Build.0 = Debug|Any CPU
+		{0DCEB08E-BA94-4968-AD0D-9504ACAA951F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{0DCEB08E-BA94-4968-AD0D-9504ACAA951F}.Debug|x86.Build.0 = Debug|Any CPU
+		{0DCEB08E-BA94-4968-AD0D-9504ACAA951F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0DCEB08E-BA94-4968-AD0D-9504ACAA951F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0DCEB08E-BA94-4968-AD0D-9504ACAA951F}.Release|x64.ActiveCfg = Release|Any CPU
+		{0DCEB08E-BA94-4968-AD0D-9504ACAA951F}.Release|x64.Build.0 = Release|Any CPU
+		{0DCEB08E-BA94-4968-AD0D-9504ACAA951F}.Release|x86.ActiveCfg = Release|Any CPU
+		{0DCEB08E-BA94-4968-AD0D-9504ACAA951F}.Release|x86.Build.0 = Release|Any CPU
+		{7282D897-351B-4F69-BF88-CDA3FC85E374}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7282D897-351B-4F69-BF88-CDA3FC85E374}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7282D897-351B-4F69-BF88-CDA3FC85E374}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{7282D897-351B-4F69-BF88-CDA3FC85E374}.Debug|x64.Build.0 = Debug|Any CPU
+		{7282D897-351B-4F69-BF88-CDA3FC85E374}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{7282D897-351B-4F69-BF88-CDA3FC85E374}.Debug|x86.Build.0 = Debug|Any CPU
+		{7282D897-351B-4F69-BF88-CDA3FC85E374}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7282D897-351B-4F69-BF88-CDA3FC85E374}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7282D897-351B-4F69-BF88-CDA3FC85E374}.Release|x64.ActiveCfg = Release|Any CPU
+		{7282D897-351B-4F69-BF88-CDA3FC85E374}.Release|x64.Build.0 = Release|Any CPU
+		{7282D897-351B-4F69-BF88-CDA3FC85E374}.Release|x86.ActiveCfg = Release|Any CPU
+		{7282D897-351B-4F69-BF88-CDA3FC85E374}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
