--- src/das-employer-favourites.ApplicationServices/Configuration/ConnectionStrings.cs
@@ -1,8 +1,0 @@
-namespace DfE.EmployerFavourites.ApplicationServices.Configuration
-{
-    public class ConnectionStrings
-    {
-        public string Redis { get; set; }
-        public string AzureStorage { get; set; }
-    }
-}
