--- src/das-employer-favourites.ApplicationServices/Domain/IEmployerAccountRepository.cs
@@ -1,9 +1,0 @@
-using System.Threading.Tasks;
-
-namespace DfE.EmployerFavourites.ApplicationServices.Domain
-{
-    public interface IEmployerAccountRepository
-    {
-        Task<string> GetEmployerAccountId(string userId);
-    }
-}
