--- src/das-employer-favourites.ApplicationServices/Domain/IFavouritesWriteRepository.cs
@@ -1,10 +1,0 @@
-using System.Threading.Tasks;
-using DfE.EmployerFavourites.ApplicationServices.Domain.WriteModel;
-
-namespace DfE.EmployerFavourites.ApplicationServices.Domain
-{
-    public interface IFavouritesWriteRepository
-    {
-        Task SaveApprenticeshipFavourites(string employerAccountId, ApprenticeshipFavourites apprenticeshipFavourite);
-    }
-}
