--- src/das-employer-favourites.ApplicationServices/Domain/ReadModel/ApprenticeshipFavourites.cs
@@ -1,8 +1,0 @@
-using System.Collections.Generic;
-
-namespace DfE.EmployerFavourites.ApplicationServices.Domain.ReadModel
-{
-    public class ApprenticeshipFavourites : List<ApprenticeshipFavourite>
-    {
-    }
-}
