--- src/das-employer-favourites.ApplicationServices/Domain/ReadModel/Provider.cs
@@ -1,8 +1,0 @@
-﻿namespace DfE.EmployerFavourites.ApplicationServices.Domain.ReadModel
-{
-    public class Provider
-    {
-        public int Ukprn { get; set; }
-        public string Name { get; set; }
-    }
-}
