--- src/das-employer-favourites.ApplicationServices/Domain/WriteModel/ApprenticeshipFavourite.cs
@@ -1,15 +1,0 @@
-using System.Collections.Generic;
-
-namespace DfE.EmployerFavourites.ApplicationServices.Domain.WriteModel
-{
-    public class ApprenticeshipFavourite
-    {
-        public ApprenticeshipFavourite()
-        {
-            Ukprns = new List<int>();
-        }
-
-        public string ApprenticeshipId { get; set; }
-        public IList<int> Ukprns { get; set; }
-    }
-}
