--- src/das-employer-favourites.ApplicationServices/Domain/WriteModel/ApprenticeshipFavourites.cs
@@ -1,8 +1,0 @@
-using System.Collections.Generic;
-
-namespace DfE.EmployerFavourites.ApplicationServices.Domain.WriteModel
-{
-    public class ApprenticeshipFavourites : List<ApprenticeshipFavourite>
-    {
-    }
-}
