--- src/das-employer-favourites.ApplicationServices/EmployerFavourites.ApplicationServices.csproj
@@ -1,23 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <RootNamespace>DfE.EmployerFavourites.ApplicationServices</RootNamespace>
-    <AssemblyName>DfE.EmployerFavourites.ApplicationServices</AssemblyName>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="NLog.Web.AspNetCore" Version="4.8.0" />
-    <PackageReference Include="SFA.DAS.Apprenticeships.Api.Client" Version="0.11.174" />
-    <PackageReference Include="SFA.DAS.NLog.Targets.Redis" Version="1.2.1" />
-    <PackageReference Include="MediatR" Version="6.0.0" />
-    <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="6.0.1" />
-    <PackageReference Include="Microsoft.Azure.Cosmos.Table" Version="1.0.1" />
-    <PackageReference Include="SFA.DAS.Account.Api.Client" Version="1.6.1" />
-    <PackageReference Include="Polly" Version="7.1.0" />
-    <PackageReference Include="SFA.DAS.Providers.Api.Client" Version="0.11.174" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Domain\ReadModel\" />
-    <Folder Include="Domain\WriteModel\" />
-  </ItemGroup>
-</Project>
