--- src/das-employer-favourites.ApplicationServices/Infrastructure/Interfaces/IFatRepository.cs
@@ -1,10 +1,0 @@
-using System.Threading.Tasks;
-
-namespace DfE.EmployerFavourites.ApplicationServices.Infrastructure.Interfaces
-{
-    public interface IFatRepository
-    {
-        Task<string> GetApprenticeshipNameAsync(string apprenticeshipId);
-        Task<string> GetProviderNameAsync(int ukprn);
-    }
-}
