+++ src/web/favourites-web.sln
@@ -9,8 +9,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitTests", "UnitTests\Unit
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmployerFavouritesApi.Client", "..\api-client\src\EmployerFavouritesApi.Client\EmployerFavouritesApi.Client.csproj", "{AFBEF52E-B55F-4BF0-A775-FF35D38BD678}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmployerFavourites.ApplicationServices", "..\das-employer-favourites.ApplicationServices\EmployerFavourites.ApplicationServices.csproj", "{557B0892-9EE7-4EAB-80F1-2C5519F3E2E3}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IntegrationTests", "IntegrationTests\IntegrationTests.csproj", "{A819AECF-0A23-4511-AEB8-6946FD197793}"
 EndProject
 Global
@@ -62,18 +60,6 @@ Global
 		{AFBEF52E-B55F-4BF0-A775-FF35D38BD678}.Release|x64.Build.0 = Release|Any CPU
 		{AFBEF52E-B55F-4BF0-A775-FF35D38BD678}.Release|x86.ActiveCfg = Release|Any CPU
 		{AFBEF52E-B55F-4BF0-A775-FF35D38BD678}.Release|x86.Build.0 = Release|Any CPU
-		{557B0892-9EE7-4EAB-80F1-2C5519F3E2E3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{557B0892-9EE7-4EAB-80F1-2C5519F3E2E3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{557B0892-9EE7-4EAB-80F1-2C5519F3E2E3}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{557B0892-9EE7-4EAB-80F1-2C5519F3E2E3}.Debug|x64.Build.0 = Debug|Any CPU
-		{557B0892-9EE7-4EAB-80F1-2C5519F3E2E3}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{557B0892-9EE7-4EAB-80F1-2C5519F3E2E3}.Debug|x86.Build.0 = Debug|Any CPU
-		{557B0892-9EE7-4EAB-80F1-2C5519F3E2E3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{557B0892-9EE7-4EAB-80F1-2C5519F3E2E3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{557B0892-9EE7-4EAB-80F1-2C5519F3E2E3}.Release|x64.ActiveCfg = Release|Any CPU
-		{557B0892-9EE7-4EAB-80F1-2C5519F3E2E3}.Release|x64.Build.0 = Release|Any CPU
-		{557B0892-9EE7-4EAB-80F1-2C5519F3E2E3}.Release|x86.ActiveCfg = Release|Any CPU
-		{557B0892-9EE7-4EAB-80F1-2C5519F3E2E3}.Release|x86.Build.0 = Release|Any CPU
 		{A819AECF-0A23-4511-AEB8-6946FD197793}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A819AECF-0A23-4511-AEB8-6946FD197793}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A819AECF-0A23-4511-AEB8-6946FD197793}.Debug|x64.ActiveCfg = Debug|Any CPU
